pyproject.toml
[project]
name = "ai-audio-text-speech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.7.0",
    "google-cloud-texttospeech>=2.33.0",
    "google-genai==1.12.1",
    "openai-whisper>=20250625",
    "python-dotenv==1.1.0",
]
